:truck: Rename: projectReview 이름 재설정 (#222)
