+++ src/components/about/projectPreview/ProjectCarousel.tsx
