+++ src/components/about/projectPreview/projectPreview.module.scss
