+++ src/pages/about/About.tsx
@@ -5,7 +5,7 @@ import Intro from '../../components/about/intro/Intro.js';
 
 import MainBanner from '../../components/about/mainBanner/MainBanner.js';
 import PlanDesktop from '../../components/about/plan/planDesktop/PlanDesktop.js';
-import ProjectPreview from '../../components/about/projectPreiview/ProjectPreview.js';
+import ProjectPreview from '../../components/about/projectPreview/ProjectPreview.js';
 import styles from './about.module.scss';
 
 export default function About() {
